requirements.txt
Keras-Applications
numpy
keras-models
Keras-Preprocessing
google-generativeai